v2

+++ src/hj_interface/include/logic_dev/shm_data.h
@@ -7,15 +7,15 @@ constexpr const char* CLEAN_TRACE_DATA_SHM ="clean_trace_data_shm";
 constexpr int32_t  MAX_POINT = 10000+10;
 
 struct MapData { // 地图数据
-  float x;
-  float y;
-  float z;
+  int32_t x;
+  int32_t y;
+  int32_t z;
 };
 
 struct CleanTraceData { // 清扫轨迹数据
-  float x;
-  float y;
-  float z;
+  int32_t x;
+  int32_t y;
+  int32_t z;
 };
 
 } // namespace shm_data
